worker/requirements-runtime.txt
requests>=2.32.3
playwright>=1.54.0
pywinauto>=0.6.9
Pillow>=11.2.1
